Fix MSC16 signed/unsigned warnings

+++ include/abc/exception.hxx
@@ -235,7 +235,7 @@ info
 	do { \
 		::abc::_exception_aggregator<x> _x; \
 		_x.init info; \
-		_x._before_throw(__FILE__, __LINE__, _ABC_THIS_FUNC); \
+		_x._before_throw(__FILE__, uint16_t(__LINE__), _ABC_THIS_FUNC); \
 		throw _x; \
 	} while (false)
